# src/main/resources/assets/api-docs.yaml
swagger: '2.0'
info:
  title: Ontology Services
  description: A set of web-services related to ontology information.
  version: 0.0.1
  license:
    name: BSD
    url: 'http://opensource.org/licenses/BSD-3-Clause'
  termsOfService: 'https://github.com/DataBiosphere/consent-ontology'
schemes:
  - https

paths:

  /schemas/data-use:
    get:
      summary: Data Use Questions and Answers
      description: JSON representation of data use questions + answers for use in defining structured data use restrictions and data access requests.
      tags:
        - Data Use
      responses:
        200:
          description: The json representation for all data use questions + answers
        default:
          description: Unexpected error

  /schemas/data-use/consent/translate:
    post:
      summary: Translate Data Use JSON into a 'Consent' Use Restriction
      description: Translate a Data Use JSON to a 'Consent' Use Restriction
      tags:
        - Data Use
      parameters:
        - name: jsonString
          in: body
          description: JSON formatted Data Use object
          required: true
          schema:
            $ref: '#/definitions/DataUse'
      responses:
        200:
          description: The use restriction in consent format
        400:
          description: invalid data use schema format
        default:
          description: Unexpected error

  /schemas/data-use/dar/translate:
    post:
      summary: Translate Data Use JSON into a 'Data Access Request' Use Restriction
      description: Translate a Data Use JSON to a 'Data Access Request' Use Restriction
      tags:
        - Data Use
      parameters:
        - name: jsonString
          in: body
          description: JSON formatted Data Use object
          required: true
          schema:
            $ref: '#/definitions/DataUse'
      responses:
        200:
          description: The use restriction in data access request format
        400:
          description: invalid data use schema format
        default:
          description: Unexpected error

  /translate:
    post:
      summary: Translate Data Use JSON into plain English
      description: The Translate service will render an English-language version of a research purpose or dataset.
      produces:
        - text/plain
      parameters:
        - in: query
          name: for
          description: Either 'dataset' or 'purpose'
          required: true
          type: string
          enum: [dataset, purpose]
        - name: body
          in: body
          description: DataUse
          required: true
          schema:
            $ref: '#/definitions/DataUse'
      tags:
        - Data Use
        - Translate
      responses:
        200:
          description: Successful translation.
        default:
          description: Unexpected error
  /translate/summary:
    post:
      summary: Translate Data Use JSON into Data Use Summary JSON
      description: The Translate service will render an structured summary with consent codes
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: DataUse
          required: true
          schema:
            $ref: '#/definitions/DataUse'
      tags:
        - Data Use
        - Translate
      responses:
        200:
          description: Successful translation.
          schema:
            $ref: '#/definitions/DataUseSummary'
        default:
          description: Unexpected error
  /translate/paragraph:
    post:
      summary: Translates Paragraph to ontology recommendations
      description: The Translate service for paragraphs will render an ontology recommendation for a given paragraph.
      parameters:
        - name: body
          in: body
          description: Paragraph to translate
          required: true
          schema:
            $ref: '#/definitions/TranslateParagraph'
      produces:
        - application/json
      tags:
        - Translate
      responses:
        200:
          description: Successful translation.
          schema:
            $ref: '#/definitions/DataUseRecommendation'
        default:
          description: Unexpected error

  /match:
    post:
      summary: Match UseRestriction Consent and Research Purpose
      description: The Match service will render a response of true or false between a Consent and a Research Purpose.
      parameters:
        - name: body
          in: body
          description: Consent and Purpose Use Restrictions.
          required: true
          schema:
            $ref: '#/definitions/MatchPair'
      tags:
        - Match
      responses:
        200:
          description: Match Result.
          schema:
            $ref: '#/definitions/MatchPairResult'
        500:
          description: Internal Server Error

  /match/v1:
    post:
      summary: Match UseRestriction Consent and Research Purpose
      description: The Match service will render a response of true or false between a Consent and a Research Purpose.
      parameters:
        - name: body
          in: body
          description: Consent and Purpose Use Restrictions.
          required: true
          schema:
            $ref: '#/definitions/MatchPair'
      tags:
        - Match
      responses:
        200:
          description: Match Result.
          schema:
            $ref: '#/definitions/MatchPairResult'
        500:
          description: Internal Server Error

  /match/v2:
    post:
      summary: Match DataUse Purpose and Consent
      description: Determine if a Research Purpose and Consent logically match.
      parameters:
        - name: body
          in: body
          description: Research Purpose and Dataset DataUse pair.
          required: true
          schema:
            $ref: '#/definitions/DataUseMatchPair'
      tags:
        - Match
      responses:
        200:
          description: Successful matching.
          schema:
            $ref: '#/definitions/DataUseMatchPairResult'
        400:
          description: Bad Request if either purpose or consent are not provided.
        500:
          description: Internal Server Error

  /validate/userestriction:
      post:
        summary: validate
        description: The Validate service will check If the posted Use Restriction has the correct format and content.
        parameters:
          - name: body
            in: body
            description: Restriction
            required: true
            schema:
              $ref: '#/definitions/Restriction'
        tags:
          - Validate
        responses:
          200:
            description: Successful validation. True for valid restriction and False for invalid.
          default:
            description: Unexpected error

  /autocomplete:
    get:
      summary: autocomplete
      description: Ontologies can be queried based on term; type and count. This provides an easy filtering mechanism that allows users to select ontology terms that are applicable to consents or research purposes.
      parameters:
        - name: q
          in: query
          description: The query term (word fragment) which the service should try and complete.
          required: true
          type: string
        - name: types
          in: query
          description: An optional list (comma-separated) of term types; if specified only those types will be searched for autocomplete suggestions. If left unspecified all available types will be searched. Available values for this service will initially be disease and organization.
          required: false
          type: string
        - name: count
          in: query
          description: An optional limit on the number of autosuggested results that are returned. Default value is 20.
          required: false
          type: number
      tags:
        - Autocomplete
      responses:
        200:
          description: Array of results.
        default:
          description: Unexpected error

  /search:
    get:
      summary: search
      description: Find ontologies based on the id.
      parameters:
        - name: id
          in: query
          description: The ontology id which the service us to find it (e.g. http://purl.obolibrary.org/obo/DOID_162).
          required: true
          type: string
      tags:
        - Search
      responses:
        200:
          description: The given id was found. Term result with parent information.
        400:
          description: Bad request.
        404:
          description: The given id was not found.
        default:
          description: Unexpected error

  /status:
    get:
      summary: System Health Status
      description: A detailed description of the various subsystem statuses that Ontology relies upon.
      tags:
        - Status
      responses:
        200:
          description: All systems are OK
        500:
          description: Some number of subsystems are not OK.
  /status/error:
    get:
      summary: Generate a logged error
      description: Generate a logged error
      parameters:
        - in: query
          name: message
          description: Message to log as an error
          required: false
          type: string
      tags:
        - Status
      responses:
        200:
          description: Error successfully logged
        500:
          description: Other error

  /liveness:
    get:
      summary: Liveness Endpoint
      description: Liveness endpoint for conducting quick health checks on the application.
      tags:
        - Status
      produces:
        - text/plain
      responses:
        200:
          description: Responds with 'Healthy' if endpoint is reachable.

  /version:
    get:
      summary: Current application version
      description: The current short hash and version of the application.
      tags:
        - Status
      responses:
        200:
          description: Successful Response
          schema:
            $ref: '#/definitions/Version'
        500:
          description: Internal Server Error

definitions:
  Restriction:
    type: object
    description: |
      A structured data use Restriction that can describe either a research purpose or a consented
      dataset. Terms are constructed according to rules defined here:
      https://github.com/DataBiosphere/consent-ontology/blob/develop/docs/UseRestrictionGrammar.md
    example:
      { "type": "named", "name": "http://purl.obolibrary.org/obo/DOID_162" }
    properties:
      type:
        type: string
        description: 'Connector (and, or)'
      operands:
        type: array
        items:
          type: string
        description: Array of ontologies.

  MatchPair:
    type: object
    properties:
      purpose:
        description: Data Use Restriction of the research purpose
        $ref: '#/definitions/Restriction'
      consent:
        description: Data Use Restriction of the consented dataset
        $ref: '#/definitions/Restriction'

  MatchPairResult:
    type: object
    properties:
      result:
        type: boolean
      matchPair:
        $ref: '#/definitions/MatchPair'

  DataUseElement:
    type: object
    description: Individual element of a DataUseSummary object
    properties:
      code:
        type: string
        description: Consent Code
      description:
        type: string
        description: Consent Code Description
    example:
      { "code": "GRU", "description": "Data is available for general research use. [GRU]"}

  DataUseSummary:
    type: object
    description: Structured summary of a DataUse object
    properties:
      primary:
        description: List of primary consent codes
        type: array
        items:
          $ref: '#/definitions/DataUseElement'
      secondary:
        type: array
        description: List of secondary consent codes
        items:
          $ref: '#/definitions/DataUseElement'
    example:
      {
        "primary": ["code": "GRU", "description": "Data is available for general research use. [GRU]"],
        "secondary": ["code": "OTHER", "description": "Data storage on the cloud is prohibited."]
      }

  DataUse:
    type: object
    description: JSON model of a set of data use questions and answers
    example:
      { "generalUse": false, "hmbResearch": true, "diseaseRestrictions": ["http://purl.obolibrary.org/obo/DOID_602"] }

  DataUseRecommendation:
    type: object
    description: JSON model of a set of data use recommendations
    example:
      { "http://purl.obolibrary.org/obo/DUO_0000042": {"title": "General Research Use", "category": "Data Use Permission" } }
  DataUseMatchPair:
    type: object
    description: JSON model of research purpose and dataset
    properties:
      purpose:
        $ref: '#/definitions/DataUse'
      consent:
        $ref: '#/definitions/DataUse'

  DataUseMatchPairResult:
    type: object
    properties:
      result:
        type: boolean
        description: Result of the match algorithm on the provided purpose and dataset
      matchPair:
        $ref: '#/definitions/DataUseMatchPair'
        description: The provided purpose and consent
      failureReasons:
        type: array
        items:
          type: string
        description: |
          Optional list of reasons why a match result may have failed. Only supported in
          in versions 2 and higher

  TranslateParagraph:
    type: object
    properties:
      paragraph:
        description: The paragraph to translate
        type: string

  Version:
    type: object
    properties:
      hash:
        type: string
        description: Current short hash of this version
      version:
        type: string
        description: Current version